문제4 풂

+++ CPP_Study/source_file/디버깅 연습 문제/Exercise_4/Knight.h
@@ -13,7 +13,7 @@ public:
 	bool IsDead();
 
 public:
-	int _hp;
+	long long _hp;
 	int _attack;
 };
